Refactor density to process and preprocess
