+++ Makefile
@@ -14,7 +14,7 @@ DEV_GENOME := "Camarosa"
 dev: | clean        ## execute with default testing arguments
 	@echo executing dev main
 	mkdir -p $(DEV_CACHE)
-	-python3 $(ROOT_DIR)/transposon/density.py $(DEV_GENES) $(DEV_TES) $(DEV_GENOME) -vv -s $(DEV_CACHE)
+	-python3 process_genome.py $(DEV_GENES) $(DEV_TES) $(DEV_GENOME) -vv -s $(DEV_CACHE)
 
 help:               ## Show this help.
 	fgrep -h "##" $(MAKEFILE_LIST) | fgrep -v fgrep | sed -e 's/\\$$//' | sed -e 's/##//'
