+++ process_genome.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 """
-Calculate transposable element density.
+Calculate transposable element density
 """
 
 __author__ = "Scott Teresi, Michael Teresi"
@@ -28,250 +28,7 @@ from transposon.verify_cache import (
     verify_gene_cache,
     revise_annotation,
 )
-from transposon.revise_annotation import Revise_Anno
-
-
-def get_nulls(my_df):
-    """
-    Print out a count of null values per column.
-    Print out the row IDs where the null values exist
-
-    Args:
-        my_df (Pandaframes): Pandaframe to check null values in
-    """
-    null_columns = my_df.columns[my_df.isnull().any()]
-    count_of_null = my_df[null_columns].isnull().sum()
-    print("Counts of null values per column: " "\n", count_of_null, "\n")
-    rows_where_null = my_df[my_df.isnull().any(axis=1)][null_columns].head()
-    print("Rows where null exist: ", "\n", rows_where_null, "\n")
-
-
-def drop_nulls(my_df, status=False):
-    """
-    Drop null values inside a Pandaframe
-
-    Args:
-        my_df (Pandaframes): Pandaframe to drop null values
-    """
-    if status:
-        print("DROPPING ROWS WITH AT LEAST ONE NULL VALUE!!!")
-    my_df = my_df.dropna(axis=0, how="any")
-    return my_df
-
-
-def swap_columns(dataframe, col_condition, col_1, col_2):
-    """
-    Swap the values of two designated columns for a row based on a column
-    condition, return the dataframe
-
-    Args:
-        my_df (Pandaframes): Pandaframe to swap columns in.
-    """
-    dataframe.loc[col_condition, [col_1, col_2]] = dataframe.loc[
-        col_condition, [col_2, col_1]
-    ].values
-    return dataframe
-
-
-def split(dataframe, group):
-    """Return list of dataframes with each element being a subset of the df.
-
-    I use this function to split by chromosome so that we may later do
-    chromosome element-wise operations.
-
-    This function is also used in revise_annotation.py to split on transposon
-    identities.
-    """
-
-    grouped_df = dataframe.groupby(group)
-    return [grouped_df.get_group(x) for x in grouped_df.groups]
-
-
-def check_density_shape(densities, transposon_data):
-    """Checks to make sure the density output is of the same dimension as the
-    transposon_data input.
-
-    Args:
-        densities (numpy.ndarray):
-            density calculations, returned from rho functions.
-        transposon_data (TransposonData):
-            transposon container
-    """
-    if densities.shape != transposon_data.starts.shape:
-        msg = "Density dataframe shape not the same size as the TE dataframe"
-        logger.critical(msg)
-        raise ValueError(msg)
-
-
-def validate_window(left_window_start, left_window_stop, window_length):
-    """
-    Validate the window specifically to make sure it doesn't go into the
-    negative values. Function invoked to validate the left-hand window.
-
-    Args:
-        window_start (int): integer value for the left window start value, we need
-        to make sure it isn't negative.
-
-        TODO add description
-
-        window_length (int)
-    """
-    if left_window_start < 0:
-        msg = "window_start is not 0 or a positive value"
-        logger.critical(msg)
-        raise ValueError(msg)
-    if left_window_start == 0:
-        window_length = left_window_stop - left_window_start + 1
-    return window_length
-
-
-def rho_left_window(gene_data, gene_name, transposon_data, window):
-    """Density to the left (downstream) of a gene.
-    When TE is between gene and window
-
-    Relevant things: No. TE bases / window
-    GeneStart so that we may calculate the WindowStart
-    WindowStart is unique to each gene. Calculated via Gstart - Window
-    WindowStart is the leftmost window
-
-    Args:
-        window (int): integer value of the current window
-        gene_data (transponson.data.GeneData): gene container
-        gene_name (hashable): name of gene to use
-        transposon_data (transponson.data.TransposonData): transposon container
-    """
-    transposon_data.check_shape()
-    gene_datum = gene_data.get_gene(gene_name)
-    g_start, g_stop, g_length = gene_datum.start_stop_len
-
-    # Define windows
-    win_length = gene_datum.win_length(window)
-    win_start = gene_datum.left_win_start(window)
-    win_stop = gene_datum.left_win_stop
-    win_length = validate_window(win_start, g_start, win_length)
-
-    # Set bounds and perform density calculation
-    lower_bound = np.maximum(win_start, transposon_data.starts)
-    upper_bound = np.minimum(win_stop, transposon_data.stops)
-    te_overlaps = np.maximum(0, (upper_bound - lower_bound + 1))
-    densities = np.divide(
-        te_overlaps, win_length, out=np.zeros_like(te_overlaps, dtype="float")
-    )
-    check_density_shape(densities, transposon_data)
-    return densities
-
-
-def rho_intra(gene_data, gene_name, transposon_data):
-    """Intra density for one gene wrt transposable elements.
-
-    The relevant ares is the gene for intra density.
-
-    Args:
-        gene_data (transponson.data.GeneData): gene container
-        gene_name (hashable): name of gene to use
-        transposon_data (transponson.data.TransposonData): transposon container
-    """
-    transposon_data.check_shape()
-    gene_datum = gene_data.get_gene(gene_name)
-    g_start, g_stop, g_length = gene_datum.start_stop_len
-    lower = np.minimum(g_stop, transposon_data.stops)
-    upper = np.maximum(g_start, transposon_data.starts)
-    te_overlaps = np.maximum(0, (lower - upper + 1))
-
-    densities = np.divide(
-        te_overlaps,
-        g_length,
-        out=np.zeros_like(te_overlaps, dtype="float"),
-        where=g_length != 0,
-    )
-
-    check_density_shape(densities, transposon_data)
-    return densities
-
-
-def rho_right_window(gene_data, gene_name, transposon_data, window):
-    """Density to the right (upstream) of a gene.
-    When TE is between gene and window
-
-    Relevant things: No. TE bases / window
-    GeneStop so that we may calculate the WindowStop
-    WindowStop is unique to each gene. Calculated via Gstop + Window
-    WindowStop is the rightmost window
-
-    Args:
-        window (int): integer value of the current window
-        gene_data (transponson.data.GeneData): gene container
-        gene_name (hashable): name of gene to use
-        transposon_data (transponson.data.TransposonData): transposon container
-    """
-    transposon_data.check_shape()
-    gene_datum = gene_data.get_gene(gene_name)
-    g_start, g_stop, g_length = gene_datum.start_stop_len
-
-    # Define windows
-    win_length = gene_datum.win_length(window)
-    win_start = gene_datum.right_win_start
-    win_stop = gene_datum.right_win_stop(window)
-
-    # Set bounds and perform density calculation
-    lower_bound = np.maximum(win_start, transposon_data.starts)
-    upper_bound = np.minimum(win_stop, transposon_data.stops)
-    te_overlaps = np.maximum(0, (upper_bound - lower_bound + 1))
-    densities = np.divide(
-        te_overlaps, win_length, out=np.zeros_like(te_overlaps, dtype="float")
-    )
-    check_density_shape(densities, transposon_data)
-    return densities
-
-
-def init_empty_densities(my_genes, my_tes, window):
-    """Initializes all of the empty columns we need in the gene file. """
-
-    # NOTE This function is a candidate for deletion
-    order_list = my_tes.Order.unique()
-    superfamily_list = my_tes.SuperFamily.unique()
-    directions = ["_downstream", "_intra", "_upstream"]
-    # left, center, right
-
-    for an_order in order_list:
-        for direction in directions:
-            col_name = str(window) + "_" + an_order + direction
-            my_genes[col_name] = np.nan
-
-    for a_superfamily in superfamily_list:
-        for direction in directions:
-            col_name = str(window) + "_" + a_superfamily + direction
-            my_genes[col_name] = np.nan
-    my_genes["TEs_inside"] = np.nan
-    return my_genes
-
-
-def check_groupings(grouped_genes, grouped_TEs, logger, genome_id):
-    """Validates the gene / TE pairs.
-
-    This is just to make sure that each pair of chromosomes are right.
-    Correct subsetting would be managed by the custom split command.
-
-    Args:
-        grouped_genes (list of pandaframes): Gene dataframes separated by chromosome
-        grouped_TEs (list of pandaframes): TE dataframes separated by chromosome
-        genome_id (str) a string of the genome name.
-    """
-    for g_element, t_element in zip(grouped_genes, grouped_TEs):
-        # print(g_element.Chromosome.iloc[:].values[0])
-        if (
-            g_element.Chromosome.iloc[:].values[0]
-            != t_element.Chromosome.iloc[:].values[0]
-        ):
-            msg = "Chromosomes do not match for the grouped_genes or grouped_TEs"
-            logger.critical(msg)
-            raise ValueError(msg)
-        try:
-            sub_gene = GeneData(g_element, genome_id)
-            subgene_uid = sub_gene.chromosome_unique_id
-        except RuntimeError as r_err:
-            logging.critical("sub gene grouping is not unique: {}".format(sub_gene))
-            raise r_err
+from preprocess import split, check_groupings
 
 
 def validate_args(args, logger):
@@ -401,22 +158,22 @@ if __name__ == "__main__":
         "--config_file",
         "-c",
         type=str,
-        default=os.path.join(path_main, "../../", "config/test_run_config.ini"),
-        help="parent path of config file",
+        default=os.path.join(path_main, "../", "config/test_run_config.ini"),
+        help="Parent path of config file",
     )
     parser.add_argument(
         "--overlap_dir",
         "-s",
         type=str,
         default=os.path.abspath("/tmp"),
-        help="parent directory to output overlap data",
+        help="Parent directory to output temporary overlap data",
     )
     parser.add_argument(
         "--filtered_input_data",
         "-f",
         type=str,
-        default=os.path.join(path_main, "../..", "filtered_input_data"),
-        help="parent directory for cached input data",
+        default=os.path.join(path_main, "../", "filtered_input_data"),
+        help="parent directory for filtered input data",
     )
     # NOTE
     # TODO if the user sets their own filtered_input_data location, this
@@ -427,20 +184,45 @@ if __name__ == "__main__":
         "-h5",
         type=str,
         default=os.path.join(
-            path_main, "../..", "filtered_input_data/chromosome_h5_cache"
+            path_main, "../", "filtered_input_data/chromosome_h5_cache"
         ),
         help="parent directory for h5 cached chromosome input data",
     )
-    parser.add_argument("--reset_h5", action="store_true")
-    parser.add_argument("--contig_del", action="store_false")
-    parser.add_argument("--revise_anno", action="store_true")
+    parser.add_argument(
+        "--reset_h5",
+        action="store_true",
+        help="""When flag is provided resets the cache of
+                        wrapped GeneData and TransposonData input files.""",
+    )
+    parser.add_argument(
+        "--contig_del",
+        action="store_false",
+        help="""When flag is provided will not delete any
+                        entries from input gene or TE annotation that have an
+                        item in the chromosome field with 'Contig'.""",
+    )
+    parser.add_argument(
+        "--reset_revise_anno",
+        action="store_true",
+        help="""When flag is provided always re-writes the revised
+                       annotation file.""",
+    )
+    parser.add_argument(
+        "--dont_use_revise_anno",
+        action="store_true",
+        help="""When flag is provided does not use a revised
+                        annotation file at all, does not create a revised
+                        annotation at all.""",
+    )
 
+    # TODO make this less dependent on the filtered_input_data arg
+    # Ask Michael, I am not sure the best way to do that.
     parser.add_argument(
         "--revised_input_data",
         "-r",
         type=str,
         default=os.path.join(
-            path_main, "../..", "filtered_input_data/revised_input_data"
+            path_main, "../", "filtered_input_data/revised_input_data"
         ),
         help="parent directory for cached revised input data",
     )
@@ -448,7 +230,7 @@ if __name__ == "__main__":
         "--output_dir",
         "-o",
         type=str,
-        default=os.path.join(path_main, "../..", "results"),
+        default=os.path.join(path_main, "../", "results"),
         help="parent directory to output results",
     )
     parser.add_argument(
@@ -521,7 +303,8 @@ if __name__ == "__main__":
     # Revise the TE annotation
     te_data_unwrapped = revise_annotation(
         te_data_unwrapped,
-        args.revise_anno,
+        args.reset_revise_anno,
+        args.dont_use_revise_anno,
         revised_transposons,
         args.revised_input_data,
         logger,
