+++ transposon/__init__.py
@@ -1,4 +1,3 @@
-
 """
 Sundry helper functions.
 """
@@ -8,8 +7,9 @@ import errno
 from functools import partial
 from os import sysconf, strerror
 import h5py
+from configparser import ConfigParser
 
-MAX_SYSTEM_RAM_GB = sysconf('SC_PAGE_SIZE') * sysconf('SC_PHYS_PAGES')/(1024.**3)
+MAX_SYSTEM_RAM_GB = sysconf("SC_PAGE_SIZE") * sysconf("SC_PHYS_PAGES") / (1024.0 ** 3)
 FILE_DNE = partial(FileNotFoundError, errno.ENOENT, strerror(errno.ENOENT))
 
 
@@ -19,8 +19,8 @@ def check_ram(ram_bytes, logger):
     if ram_bytes < 0:
         logger.critical("cache %i bytes < 0" % ram_bytes)
         raise ValueError()
-    elif ram_bytes/(1024.**3) > MAX_SYSTEM_RAM_GB:
-        ram_gb = ram_bytes/(1024.**3)
+    elif ram_bytes / (1024.0 ** 3) > MAX_SYSTEM_RAM_GB:
+        ram_gb = ram_bytes / (1024.0 ** 3)
         msg = "cache %i GB > system %i GB" % ram_gb
         logger.critical(msg)
         raise ValueError(msg)
@@ -37,8 +37,7 @@ def write_vlen_str_h5py(h5file, strings, dataset_key):
 
     vlen = h5py.special_dtype(vlen=str)
     n_strings = sum(1 for s in strings)
-    dset = h5file.create_dataset(
-        dataset_key, (n_strings,), dtype=vlen)
+    dset = h5file.create_dataset(dataset_key, (n_strings,), dtype=vlen)
     dset[:] = strings
 
 
@@ -51,3 +50,19 @@ def read_vlen_str_h5py(h5file, dataset_key):
     """
 
     return h5file[dataset_key][:].tolist()
+
+
+def read_density_config(filepath):
+
+    # TODO validate filepath
+    parser = ConfigParser()
+    parser.read(filepath)
+    first_window_size = parser.getint("density_parameters", "first_window_size")
+    window_delta = parser.getint("density_parameters", "window_delta")
+    last_window_size = parser.getint("density_parameters", "last_window_size")
+    alg_parameters = {
+        "first_window_size": first_window_size,
+        "window_delta": window_delta,
+        "last_window_size": last_window_size,
+    }
+    return alg_parameters
