+++ transposon/transposon_data.py
@@ -32,7 +32,7 @@ class TransposonData(object):
         Args:
             transposable_elements (pandas.DataFrame): transposable element data frame.
             genome_id (str): a string of the genome name, provided via arg
-            parser in density.py
+            parser in process.py
         """
 
         self._logger = logger or logging.getLogger(__name__)
@@ -47,9 +47,12 @@ class TransposonData(object):
         self.genome_id = None
 
     @classmethod
-    def mock(cls,
-             start_stop=np.array([[0, 9], [10, 19], [20, 29]]), chromosome='Chr_ID',
-             genome_id="fake_genome_id"):
+    def mock(
+        cls,
+        start_stop=np.array([[0, 9], [10, 19], [20, 29]]),
+        chromosome="Chr_ID",
+        genome_id="fake_genome_id",
+    ):
         """Mocked data for testing.
 
         Args:
@@ -60,7 +63,7 @@ class TransposonData(object):
         data = []
         family = "Family_0"  # FUTURE may want to parametrize family name later
         # NB overall order is not important but the names are
-        columns = ['Start', 'Stop', 'Length', 'Order', 'SuperFamily', 'Chromosome']
+        columns = ["Start", "Stop", "Length", "Order", "SuperFamily", "Chromosome"]
         for gi in range(n_genes):
             g0 = start_stop[gi, 0]
             g1 = start_stop[gi, 1]
@@ -74,18 +77,17 @@ class TransposonData(object):
         frame = pd.DataFrame(data, columns=columns)
         return TransposonData(frame, genome_id)
 
-
-    def write(self, filename, key='default'):
+    def write(self, filename, key="default"):
         """Write a Pandaframe to disk.
 
         Args:
             filename (str): a string of the filename to write.
             key (str): identifier for the group (dataset) in the hdf5 obj.
         """
-        self.data_frame.to_hdf(filename, key=key, mode='w')
+        self.data_frame.to_hdf(filename, key=key, mode="w")
 
     @classmethod
-    def read(cls, filename, genome_id, key='default'):
+    def read(cls, filename, genome_id, key="default"):
         """Read from disk. Returns a wrapped Pandaframe from an hdf5 file
 
         Args:
@@ -105,7 +107,7 @@ class TransposonData(object):
             filename (str): a string of the filename to write.
         """
 
-        self.data_frame.to_csv(filename, header=True, index=False, sep='\t')
+        self.data_frame.to_csv(filename, header=True, index=False, sep="\t")
 
     def write_gff(self, filename):
         """
@@ -116,21 +118,33 @@ class TransposonData(object):
         """
         self.gff_data_frame = self.data_frame.copy(deep=True)
         self.gff_data_frame.rename(columns={"Stop": "End"}, inplace=True)
-        self.gff_data_frame['Source'] = 'Test'
-        self.gff_data_frame['Score'] = '.'
-        self.gff_data_frame['Phase'] = '1'
-        self.gff_data_frame['Attributes'] = 'Test'
-        self.gff_data_frame['Feature'] = self.gff_data_frame['Order'] + '/' + self.gff_data_frame['SuperFamily']
+        self.gff_data_frame["Source"] = "Test"
+        self.gff_data_frame["Score"] = "."
+        self.gff_data_frame["Phase"] = "1"
+        self.gff_data_frame["Attributes"] = "Test"
+        self.gff_data_frame["Feature"] = (
+            self.gff_data_frame["Order"] + "/" + self.gff_data_frame["SuperFamily"]
+        )
 
-        self.gff_data_frame = self.gff_data_frame.astype({"Start": int, "End":
-                                                          int, 'Phase': int})
+        self.gff_data_frame = self.gff_data_frame.astype(
+            {"Start": int, "End": int, "Phase": int}
+        )
 
-        self.gff_data_frame = self.gff_data_frame[['Chromosome', 'Source',
-                                                   'Feature', 'Start', 'End',
-                                                   'Score', 'Strand', 'Phase',
-                                                   'Attributes']]
-        self.gff_data_frame = self.gff_data_frame.sort_values(by=['Start'])
-        self.gff_data_frame.to_csv(filename, header=False, index=False, sep='\t')
+        self.gff_data_frame = self.gff_data_frame[
+            [
+                "Chromosome",
+                "Source",
+                "Feature",
+                "Start",
+                "End",
+                "Score",
+                "Strand",
+                "Phase",
+                "Attributes",
+            ]
+        ]
+        self.gff_data_frame = self.gff_data_frame.sort_values(by=["Start"])
+        self.gff_data_frame.to_csv(filename, header=False, index=False, sep="\t")
 
     @property
     def number_elements(self):
@@ -145,8 +159,7 @@ class TransposonData(object):
         at a time
         """
         pass
-        #for superfamily in self.superfamily_name_set:
-
+        # for superfamily in self.superfamily_name_set:
 
     def check_shape(self):
         """Checks to make sure the columns of the TE data are the same size.
@@ -158,16 +171,18 @@ class TransposonData(object):
         start = self.starts.shape
         stop = self.stops.shape
         if start != stop:
-            msg = ("Input TE missing fields: starts.shape {}  != stops.shape {}"
-                   .format(start, stop))
+            msg = "Input TE missing fields: starts.shape {}  != stops.shape {}".format(
+                start, stop
+            )
             self._logger.critical(msg)
             raise ValueError(msg)
 
         # TODO verify that this isn't weird
         length = self.lengths.shape
         if start != length:
-            msg = ("Input TE missing fields: starts.shape {}  != lengths.shape {}"
-                   .format(start, stop))
+            msg = "Input TE missing fields: starts.shape {}  != lengths.shape {}".format(
+                start, stop
+            )
             self._logger.critical(msg)
             raise ValueError(msg)
 
@@ -179,9 +194,8 @@ class TransposonData(object):
 
     def add_genome_id(self, genome_id):
         """Add the genome_id as an extra column to the gene_dataframe"""
-        self.data_frame.loc[:,'Genome_ID'] = genome_id
+        self.data_frame.loc[:, "Genome_ID"] = genome_id
         self.genome_id = genome_id
-
 
     @property
     def genome_unique_id(self):
@@ -199,8 +213,7 @@ class TransposonData(object):
         if not genome_id_list:
             raise RuntimeError("column 'Geneome_ID' is empty")
         elif len(genome_id_list) > 1:
-            raise RuntimeError("Genome IDs are are not unique: %s" %
-                               genome_id_list)
+            raise RuntimeError("Genome IDs are are not unique: %s" % genome_id_list)
         else:
             return genome_id_list[0]  # MAGIC NUMBER list to string
